Number sorting program.
This is a program I wrote for a python course I'm currently taking. There is a slight issue with the code where when inputting the largest number into the second number input, the program no longer displays what the middle number is.

I don't know if anyone will ever see this or this repo in general, but if you do, feedback is always appreciated!