Bug fix - SortNumbers.py
I ask a buddy of mine for some help with my input issue (thanks again Fabian!!), and after a little bit more testing we found what the issue was. When the userInput3 was less than both userInput1 and userInput2, the program would skip right over the middle section of code and only output the largest and the smallest values. He did a little more research and testing on his end, and we came to the conclusion that what needed to be done was just adding a secondary if-else to check through the inputs again and confirm the middle input. After adding this, the program works just fine with any of the numbers being in any of the inputs!